cgraph agxbuf: support storing short strings inline
This commit implements Small String Optimization (SSO) for `axgbuf`. Strings
up to a certain size (23 bytes unterminated on x86-64) can be stored inline
within an `agxbuf` structure itself with no external backing memory.

There is currently no way to create a buffer with `AGXBUF_INLINE`. Neither
`agxbuf_init` nor zero-initialization gives a path to this. The ability to
create inline strings will be added in an upcoming commit.

The expansion logic in `agxbmore` is adjusted in a way that does not preserve
the exact previous behavior but aims to preserve its intent. That is, expansions
from an initially small buffer size will round up to `BUFSIZ` (typically 128).

When a string is stored inline, its size is not maintained anywhere. The string
data is kept NUL terminated. That is, `xb->store[sizeof(xb->store) - 1]` is
always `'\0'` for inline strings.

Note that implementing `agxbuse` on this type of `agxbuf` required a bit of
lateral thinking.