cgraph agxbuf: make inline strings the default
This change switches both `agxbinit` and zero-initialization (`agxbuf xb = {0}`)
to result in a buffer with `AGXBUF_INLINE` set. This means string data is
written inline until it becomes too large to store within the `agxbuf` struct
itself, when it is then relocated to the heap.

This is an optimization. Short dynamic strings can now be written completely
without heap allocation. For example, stringifying a number
(`agxbprint(&xb, "%d", i)`) will fit fully within the inline buffer.