cgraph agxbuf: increase available inline storage
The `located` member in this struct determines where the backing store is. When
storage is inline, its value is 0. Its only possible values are 0, 1, and 2. It
is also the last member in this struct. Putting these pieces of information
together, we can move it to the last _byte_ of this struct allowing extra room
for inline storage. On top of this, we can claim a further byte by noting that
the inline string is naturally NUL terminated by `AGXBUF_INLINE` itself being
present in the last byte.

This further optimization on top of the previous SSO rearrangements gains us an
extra 8 bytes on x86-64. That is, strings in the range 24 - 31 bytes that
previously would have required heap allocation can now be stored inline just
like strings <24 bytes.