common: define font name aliases as a list in estimation LUT
There is no need to tokenize these font lists at runtime when we can do it at
compile time. This makes the intent of this code clearer to both humans and the
compiler. It also allows future additions to this table of font names containing
the pipe character, though these are ignored during comparison anyway.